refactor: move IsMutuallyAssignable to unstable-types
